Create Idea
I was interviewing with a company and was tasked to create aspects of Minesweeper. It would accept inputs in the form (3, [2, 1, 0, 2], [0, 2, 1, 2]). The first number is the length and width of the square. The second input is the row coordinates for the bombs and the third input is the coordinates in the columns. This code is the basis for my game.